split up test_agents into different files
