+++ tests/test_deep/__init__.py
@@ -1,3 +1,13 @@
+from tests.test_deep.test_agents import (
+    test_a2c,
+    test_ddpg,
+    test_dqn,
+    test_dqn_cnn,
+    test_ppo1,
+    test_sac,
+    test_td3,
+    test_vpg,
+)
 from tests.test_deep.test_common import (
     TestUtils,
     test_loggers,
