+++ tests/test_deep/test_agents/__init__.py
@@ -1,0 +1,8 @@
+from tests.test_deep.test_agents.test_a2c import test_a2c
+from tests.test_deep.test_agents.test_ddpg import test_ddpg
+from tests.test_deep.test_agents.test_dqn import test_dqn
+from tests.test_deep.test_agents.test_dqn_cnn import test_dqn_cnn
+from tests.test_deep.test_agents.test_ppo1 import test_ppo1
+from tests.test_deep.test_agents.test_sac import test_sac
+from tests.test_deep.test_agents.test_td3 import test_td3
+from tests.test_deep.test_agents.test_vpg import test_vpg
